api/requirements.txt
fastapi
uvicorn
pymongo
mysql-connector-python
python-dotenv
aiofiles
pandas
numpy
celery
python-multipart
